Apply currency normalization to payout ticker too
(cherry picked from commit 75a9383fb8b06fce0f99baa2d8bd8bc2c3ee4a1b)
