+++ src/check_payout_address.c
@@ -10,6 +10,7 @@
 #include "parse_coin_config.h"
 #include "printable_amount.h"
 #include "menu.h"
+#include "ticker_normalization.h"
 
 int check_payout_address(swap_app_context_t *ctx, const command_t *cmd, SendFunction send) {
     static buf_t config;
@@ -60,10 +61,8 @@ int check_payout_address(swap_app_context_t *ctx, const command_t *cmd, SendFunc
         return reply_error(ctx, INCORRECT_COMMAND_DATA, send);
     }
 
-    // Check that given ticker match current context
-    if (strlen(ctx->received_transaction.currency_to) != ticker.size ||
-        strncmp(ctx->received_transaction.currency_to, (const char *) ticker.bytes, ticker.size) !=
-            0) {
+    // Check that refund ticker matches the current context
+    if (!check_matching_ticker(&ticker, ctx->received_transaction.currency_to)) {
         PRINTF("Error: Payout ticker doesn't match configuration ticker\n");
 
         return reply_error(ctx, INCORRECT_COMMAND_DATA, send);
