+++ src/check_refund_address.c
@@ -10,43 +10,7 @@
 #include "menu.h"
 #include "process_transaction.h"
 #include "parse_coin_config.h"
-
-// Check if a given ticker matches the current swap context
-static bool check_matching_ticker(swap_app_context_t *ctx, const buf_t *ticker) {
-    char normalized_ticker_name[9];
-    uint8_t normalized_ticker_len;
-
-    // Normalize the ticker name first
-    memcpy(normalized_ticker_name, ticker->bytes, sizeof(normalized_ticker_name));
-    normalized_ticker_name[ticker->size] = '\0';
-    to_uppercase(normalized_ticker_name, ticker->size);
-    set_ledger_currency_name(normalized_ticker_name, sizeof(normalized_ticker_name));
-    // Recalculate length in case it changed
-    normalized_ticker_len = strlen(normalized_ticker_name);
-
-    if (strncmp(normalized_ticker_name,
-                (const char *) ticker->bytes,
-                MAX(normalized_ticker_len, ticker->size)) != 0) {
-        PRINTF("Normalized ticker, from '%.*s' to '%s'\n",
-               ticker->size,
-               ticker->bytes,
-               normalized_ticker_name);
-    }
-
-    if (strlen(ctx->received_transaction.currency_from) != normalized_ticker_len ||
-        strncmp(ctx->received_transaction.currency_from,
-                normalized_ticker_name,
-                normalized_ticker_len) != 0) {
-        PRINTF("Error: Refund ticker '%.*s' doesn't match expected ticker '%.*s'\n",
-               normalized_ticker_len,
-               normalized_ticker_name,
-               strlen(ctx->received_transaction.currency_from),
-               ctx->received_transaction.currency_from);
-        return false;
-    }
-
-    return true;
-}
+#include "ticker_normalization.h"
 
 int check_refund_address(swap_app_context_t *ctx, const command_t *cmd, SendFunction send) {
     static buf_t config;
@@ -96,7 +60,9 @@ int check_refund_address(swap_app_context_t *ctx, const command_t *cmd, SendFunc
     }
 
     // Check that refund ticker matches the current context
-    if (!check_matching_ticker(ctx, &ticker)) {
+    if (!check_matching_ticker(&ticker, ctx->received_transaction.currency_from)) {
+        PRINTF("Error: Refund ticker doesn't match configuration ticker\n");
+
         return reply_error(ctx, INCORRECT_COMMAND_DATA, send);
     }
